+++ apps/mobile/src/components/common/Container.tsx
@@ -1,3 +1,7 @@
+import { ALLOWED_HEX_CHARACTERS } from '@lenstube/constants'
+import type { Profile } from '@lenstube/lens'
+import { useUserProfilesQuery } from '@lenstube/lens'
+import { useWalletConnectModal } from '@walletconnect/modal-react-native'
 import { LinearGradient } from 'expo-linear-gradient'
 import type { FC, PropsWithChildren } from 'react'
 import React from 'react'
@@ -5,6 +9,7 @@ import { StyleSheet } from 'react-native'
 
 import { theme } from '~/helpers/theme'
 import useMobileStore from '~/store'
+import { useMobilePersistStore } from '~/store/persist'
 
 const styles = StyleSheet.create({
   background: {
@@ -12,26 +17,49 @@ const styles = StyleSheet.create({
   }
 })
 
-const hexCharacters = [
-  0,
-  1,
-  2,
-  3,
-  4,
-  5,
-  6,
-  7,
-  8,
-  9,
-  'A',
-  'B',
-  'C',
-  'D',
-  'E',
-  'F'
-]
-
 const Container: FC<PropsWithChildren> = ({ children }) => {
+  const { address } = useWalletConnectModal()
+
+  const setChannels = useMobileStore((state) => state.setChannels)
+  const setSelectedChannel = useMobileStore((state) => state.setSelectedChannel)
+  const setUserSigNonce = useMobileStore((state) => state.setUserSigNonce)
+  const setSelectedChannelId = useMobilePersistStore(
+    (state) => state.setSelectedChannelId
+  )
+  const selectedChannelId = useMobilePersistStore(
+    (state) => state.selectedChannelId
+  )
+
+  const setUserChannels = (channels: Profile[]) => {
+    setChannels(channels)
+    const channel = channels.find((ch) => ch.id === selectedChannelId)
+    setSelectedChannel(channel ?? channels[0])
+    setSelectedChannelId(channel?.id)
+  }
+
+  const resetAuthState = () => {
+    setSelectedChannel(null)
+    setSelectedChannelId(null)
+  }
+
+  useUserProfilesQuery({
+    variables: {
+      request: { ownedBy: [address] }
+    },
+    skip: !address,
+    onCompleted: (data) => {
+      const channels = data?.profiles?.items as Profile[]
+      if (!channels.length) {
+        return resetAuthState()
+      }
+      setUserChannels(channels)
+      setUserSigNonce(data?.userSigNonces?.lensHubOnChainSigNonce)
+    },
+    onError: () => {
+      setSelectedChannelId(null)
+    }
+  })
+
   const homeGradientColor = useMobileStore((state) => state.homeGradientColor)
   const setHomeGradientColor = useMobileStore(
     (state) => state.setHomeGradientColor
@@ -43,8 +71,10 @@ const Container: FC<PropsWithChildren> = ({ children }) => {
     }
     let hexColorRep = '#'
     for (let index = 0; index < 6; index++) {
-      const randomPosition = Math.floor(Math.random() * hexCharacters.length)
-      hexColorRep += hexCharacters[randomPosition]
+      const randomPosition = Math.floor(
+        Math.random() * ALLOWED_HEX_CHARACTERS.length
+      )
+      hexColorRep += ALLOWED_HEX_CHARACTERS[randomPosition]
     }
     setHomeGradientColor(hexColorRep)
     return (hexColorRep += '35')
